convert DropdownMenu to ts
