--- src/DropdownMenu/index.d.ts
@@ -1,3 +1,0 @@
-import DropdownMenu from './DropdownMenu';
-import Menu from './Menu';
-export { DropdownMenu, Menu };
