--- src/DropdownMenu/index.js
@@ -1,4 +1,0 @@
-import DropdownMenu from './DropdownMenu';
-export { DropdownMenu };
-import Menu from './Menu';
-export { Menu };
