+++ src/DropdownMenu/index.ts
@@ -1,0 +1,3 @@
+import DropdownMenu, { DropdownMenuProps } from './DropdownMenu';
+import Menu, { MenuProps } from './Menu';
+export { DropdownMenu, DropdownMenuProps, Menu, MenuProps };
